refactor(portal): relocate common/{apps/fetch-apps,colors} to full-path mirror
Move the two merged flat common areas to the full-path layout
(common/layout/AppSelector/... and common/Profile/types) so all of scenes/common
mirrors the Portal subtree. Byte-identical relocation + import repoints; no
behavior change. me-query left as-is (common-native, no Portal origin).
