+++ web/scenes/Portal/Profile/page/ColorSelector/index.tsx
@@ -1,5 +1,5 @@
 import { Option } from "@/scenes/Portal/Profile/page/ColorSelector/Option";
-import { Color, colors } from "@/scenes/common/colors";
+import { Color, colors } from "@/scenes/common/Profile/types";
 
 type ColorSelectorProps = {
   value: Color;
