+++ web/scenes/common/layout/AppSelector/graphql/client/fetch-apps.gql
